requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
PyMySQL==1.1.0
SQLAlchemy==2.0.23
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
bcrypt==4.1.2
python-dotenv==1.0.0
python-dateutil==2.8.2
pytest==7.4.3
pytest-cov==4.1.0

